fix: fix pagination, garage and winners contollers
